Generate RTL kerning.
Apparently the simple kerning data from UFO is not sufficient for simple RTL kerning. We need to know if we do RTL or LTR Kerning.
This overrides the method getFeatureRulesForPairs of ufo2fdk.kernFeatureWriter with a version that writes simple RTL kerning pairs, instead of LTR. So with a file that has exclusiveley RTL-Kerning, this works out.

Thanks @khaledhosny for the help and info.

The discussion can be found here: https://groups.google.com/forum/#!msg/googlefonts-discuss/kLnaUCegGjA/rjri18RMOvwJ